+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Version 1.2
 * Added `--db-tenant` flag to specify the tenant for the vdb driver. Default options for `db-bucket`, `db-schema` and `db-table`
+* Support env variables from VDB schema (ORION-199045).
 
 ## Version 1.1
 * various bug fixes in the collector ebpf